added daily missions

+++ tti2sdk/src/main/java/com/timwe/tti2sdk/data/entity/Mission.kt
@@ -10,7 +10,7 @@ data class Mission(
     var flagText: String,
     var extraFlagColor: String,
     var extraFlagText: String,
-    var distance: Int,
+    var distance: Long,
     var distanceUnit: String,
     var title: String,
     var subtitle: String,
@@ -85,3 +85,26 @@ data class Reward (
     val imageUrl: String,
     val label: String
 )
+
+//    fun getMockedMissions(qtd: Int): ArrayList<Mission>{
+//        val list = ArrayList<Mission>()
+//        for(i in 1..qtd) {
+//            list.add(
+//                Mission(
+//                    id = i,
+//                    type = if(i%2 !=0) "hasActions" else "error",
+//                    flagColor = "red",
+//                    flagText = if(i%2 !=0) "Daily" else "Weekly",
+//                    extraFlagColor = "gold",
+//                    extraFlagText = "Actions needed",
+//                    distance = if(i%2 !=0) 3 else 9,
+//                    distanceUnit = if(i%2 !=0) "km" else "m",
+//                    title = if(i%2 !=0) "Title 1" else "Title 2",
+//                    subtitle = if(i%2 !=0) "12 Days of Christmas 12 Days of Christmas 12 Days of Christmas 12 Days of Christmas 12 Days of Christmas 12 Days of Christmas" else "12 Days of Christmas 12 Days of Christmas 12 Days of Christmas 12 Days of Christmas 12 Days of Christmas 12 Days of Christmas",
+//                    extraInfo = if(i%2 !=0) "extra informations 1" else "extra informations 2",
+//                    qtdItens = "2 of 5"
+//                )
+//            )
+//        }
+//        return list
+//    }
